Implemented increase and decrease functions for Brightness, Colour and Contrast classes

+++ src/video/V4L/Colour.h
@@ -22,7 +22,7 @@ public:
      * \param value that will be added (step value)
      * \return a bool representing the success of the operation
      */
-    virtual bool increase(int value);
+    virtual bool increase(__u16 value);
 
 
     //! Method to decrease the value of the colour
@@ -30,7 +30,7 @@ public:
      * \param value that will be subtracted (step value)
      * \return a bool representing the success of the operation
      */
-    virtual bool decrease(int value);
+    virtual bool decrease(__u16 value);
 
 
     //! Method to set the colour to an exact value
@@ -38,7 +38,7 @@ public:
      * \param value that will be added (step value)
      * \return a bool representing the success of the operation
      */
-    virtual bool setTo(int value);
+    virtual bool setTo(__u16 value);
 
 
     //! Method to reset the colour to its initial value
